🐣👹 Checkpoint ./comandos/perfil.js:9502259/159:10698757/1589 ./comandos/limpar.js:10260136/343 ./watch.json:9502259/6:10698757/4 ./comandos/login.js:9502259/67

+++ comandos/login.js
@@ -6,6 +6,7 @@ if(!args[1]) return message.reply("Faltou sua senha")
 let name = message.mentions.members.first() || args[0]
 if(!name) return message.reply("Esse Usuario nao Existe")
 await Database.Usuarios.findOne({ _id: name.id || args[0]}, function(erro, dados) {  
+if(!dados) return message.reply("Conta Inexistente")
 message.delete();
 if(dados.idlogado > 0) return message.reply("Alguem ja está nessa Conta")
 //if(dados.logado == true) return message.reply("Voce ja está Logado")
